requirements.txt
python-telegram-bot
yfinance
pandas
numpy
transformers
torch
python-dotenvzx,